requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==1.10.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Task Queue
celery==5.3.4
redis==5.0.1

# AI & ML
openai==1.3.0
elevenlabs==0.2.24
moviepy==1.0.3
Pillow==10.4.0

# Video Processing
ffmpeg-python==0.2.0

# YouTube Integration
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0

# Monitoring & Logging
prometheus-client==0.18.0
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1

# Security
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# Utilities
python-multipart==0.0.6
aiofiles==23.2.1
tenacity==8.2.3 